+++ startos/versions/index.ts
@@ -1,8 +1,9 @@
 import { VersionGraph } from '@start9labs/start-sdk'
 import { current } from './current'
 import { v0_2_0 } from './v0_2_0'
+import { v0_3_0 } from './v0_3_0'
 
 export const versionGraph = VersionGraph.of({
   current,
-  other: [v0_2_0],
+  other: [v0_2_0, v0_3_0],
 })